extended address runs partially

+++ Sparky/3rdparty/libmodbus/src/modbus.h
@@ -180,7 +180,7 @@ typedef enum
 } modbus_error_recovery_mode;
 
 typedef void (*modbus_monitor_add_item_fnc_t)(modbus_t *ctx,
-        uint8_t isOut, uint8_t slave, uint8_t func, uint16_t addr, uint16_t nb, 
+        uint8_t isOut, uint16_t slave, uint8_t func, uint16_t addr, uint16_t nb,
         uint16_t expectedCRC, uint16_t actualCRC );
 typedef void (*modbus_monitor_raw_data_fnc_t)(modbus_t *ctx,
         uint8_t *data, uint8_t dataLen, uint8_t addNewline);
